requirements.txt
PyJWT==1.6.0
psycopg2-binary
psycopg2==2.6.2
requests